Move `TimeTypes.hpp/ipp` to `Vocabulary/`
