+++ Sts1CobcSw/CobcSoftware/SpiStartupTestAndSupervisorThread.cpp
@@ -10,7 +10,7 @@
 #include <Sts1CobcSw/Periphery/Spis.hpp>
 #include <Sts1CobcSw/Utility/DebugPrint.hpp>
 #include <Sts1CobcSw/Utility/ErrorDetectionAndCorrection.hpp>
-#include <Sts1CobcSw/Utility/TimeTypes.hpp>
+#include <Sts1CobcSw/Vocabulary/TimeTypes.hpp>
 
 #include <strong_type/affine_point.hpp>
 #include <strong_type/difference.hpp>
