+++ Sts1CobcSw/CobcSoftware/TopicsAndSubscribers.hpp
@@ -2,8 +2,8 @@
 
 
 #include <Sts1CobcSw/Telemetry/TelemetryRecord.hpp>
-#include <Sts1CobcSw/Utility/TimeTypes.hpp>
 #include <Sts1CobcSw/Vocabulary/ProgramId.hpp>
+#include <Sts1CobcSw/Vocabulary/TimeTypes.hpp>
 
 #include <rodos_no_using_namespace.h>
