+++ Sts1CobcSw/Hal/HardwareSpi.hpp
@@ -3,7 +3,7 @@
 
 #include <Sts1CobcSw/Hal/Spi.hpp>
 #include <Sts1CobcSw/Utility/RodosTime.hpp>
-#include <Sts1CobcSw/Utility/TimeTypes.hpp>
+#include <Sts1CobcSw/Vocabulary/TimeTypes.hpp>
 
 #include <rodos_no_using_namespace.h>
