+++ Sts1CobcSw/Hal/SpiMock.hpp
@@ -2,7 +2,7 @@
 
 
 #include <Sts1CobcSw/Hal/Spi.hpp>
-#include <Sts1CobcSw/Utility/TimeTypes.hpp>
+#include <Sts1CobcSw/Vocabulary/TimeTypes.hpp>
 
 #include <cstddef>
 #include <cstdint>
