+++ Sts1CobcSw/Hal/Uart.hpp
@@ -2,7 +2,7 @@
 
 
 #include <Sts1CobcSw/Outcome/Outcome.hpp>
-#include <Sts1CobcSw/Utility/TimeTypes.hpp>
+#include <Sts1CobcSw/Vocabulary/TimeTypes.hpp>
 
 #include <rodos_no_using_namespace.h>
