+++ Sts1CobcSw/Periphery/Flash.hpp
@@ -6,7 +6,7 @@
 
 #include <Sts1CobcSw/Outcome/Outcome.hpp>
 #include <Sts1CobcSw/Serial/Byte.hpp>
-#include <Sts1CobcSw/Utility/TimeTypes.hpp>
+#include <Sts1CobcSw/Vocabulary/TimeTypes.hpp>
 
 #include <array>
 #include <cstddef>
