+++ Sts1CobcSw/Vocabulary/TimeTypes.hpp
@@ -50,4 +50,4 @@ template<std::endian endianness>
 }
 
 
-#include <Sts1CobcSw/Utility/TimeTypes.ipp>  // IWYU pragma: keep
+#include <Sts1CobcSw/Vocabulary/TimeTypes.ipp>  // IWYU pragma: keep
